[test-schools] Refactor testSchoolsKS to use common TestInstance and add dir-check
