+++ testing/testSchoolsKS.py
@@ -3,10 +3,8 @@ import os
 import subprocess
 import sys
 import xml.etree.ElementTree as ET
-from pathlib import Path
-from pprint import pprint
-from sys import stderr
 
+from testing.testInstance import TestInstance, test_instances
 from utility import position_on_edge
 
 if 'SUMO_HOME' in os.environ:
@@ -18,38 +16,6 @@ else:
 import sumolib
 
 
-class TestInstance:
-    def __init__(self, name: str, net_file: str, gen_stats_in_file: str, real_stats_file: str, centre: str):
-        self.name = name
-        self.net_file = net_file
-        self.gen_stats_in_file = gen_stats_in_file
-        self.real_stats_file = real_stats_file
-        self.centre = centre
-
-        try:
-            Path(self.net_file).resolve(strict=True)
-            Path(self.gen_stats_in_file).resolve(strict=True)
-            Path(self.real_stats_file).resolve(strict=True)
-        except FileNotFoundError:
-            print(f"Files for test instance: {self.name} does not exist", file=stderr)
-            pprint(self.__dict__)
-            exit(1)
-
-
-test_instances = [
-    TestInstance("aalborg", "../in/cities/aalborg.net.xml", "../in/cities/aalborg.stat.xml",
-                 "../stats/aalborg.stat.xml", "9396,12766"),
-    TestInstance("esbjerg", "../in/cities/esbjerg.net.xml", "../in/cities/esbjerg.stat.xml",
-                 "../stats/esbjerg.stat.xml", "7476,1712"),
-    TestInstance("randers", "../in/cities/randers.net.xml", "../in/cities/randers.stat.xml",
-                 "../stats/randers.stat.xml", "19516,6606"),
-    TestInstance("slagelse", "../in/cities/slagelse.net.xml", "../in/cities/slagelse.stat.xml",
-                 "../stats/slagelse.stat.xml", "6073,4445"),
-    TestInstance("vejen", "../in/cities/vejen.net.xml", "../in/cities/vejen.stat.xml", "../stats/vejen.stat.xml",
-                 "37800,3790")
-]
-
-
 def write_school_coords(net: sumolib.net.Net, stats: ET.ElementTree, filename):
     """
     Writes all schools' positions found in stats file to a csv called 'filename'. These coordinates can be used for
@@ -57,6 +23,11 @@ def write_school_coords(net: sumolib.net.Net, stats: ET.ElementTree, filename):
     :param net: network file that the schools in stats file is placed on
     :param stats: stats file parsed with ElementTree containing schools :param filename: name of csv to be written
     """
+    # Ensure that output directory exists
+    directory = "school_coordinates"
+    if not os.path.exists(directory):
+        os.mkdir(directory)
+
     xml_schools = [xml_school for xml_school in stats.find("schools").findall("school")]
 
     if xml_schools is None:
@@ -72,8 +43,8 @@ def write_school_coords(net: sumolib.net.Net, stats: ET.ElementTree, filename):
         positions.append(position_on_edge(edge, pos))
 
     # workaround to append columns to csv file. read old_csv, make a new csv and write to this, rename to old csv when done
-    old_csv = f'../out/{filename}.csv'
-    new_csv = f'../out/{filename}-new.csv'
+    old_csv = f'{directory}/{filename}-school-coords.csv'
+    new_csv = f'{directory}/{filename}-school-coords-new.csv'
 
     # if the old_csv already exists, do as mentioned above
     if os.path.exists(old_csv):
